backend/requirements.txt
fastapi
uvicorn[standard]
pytesseract
opencv-python-headless
Pillow
numpy
python-multipart
pydantic
pydantic-settings
ruff
pytest
httpx
cachetools
slowapi
python-json-logger
sentry-sdk[fastapi]
supabase
sqlalchemy
requests
beautifulsoup4
google-genai
